Code cleanup.
Code cleanup.

Code cleanup.

Code cleanup.

Code cleanup.

Code cleanup.

+++ src/main/java/net/sf/odinms/client/Pet.java
@@ -1,24 +1,25 @@
 package net.sf.odinms.client;
 
-import java.awt.Point;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
-import java.util.List;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import net.sf.odinms.database.DatabaseConnection;
 import net.sf.odinms.server.MapleItemInformationProvider;
 import net.sf.odinms.server.movement.AbsoluteLifeMovement;
 import net.sf.odinms.server.movement.LifeMovement;
 import net.sf.odinms.server.movement.LifeMovementFragment;
 
+import java.awt.*;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
+import java.util.List;
+import java.util.Optional;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
 /**
- *
  * @author Matze
  */
-public class MaplePet extends Item {
+public class Pet extends Item {
 
     private String name;
     private int uniqueid;
@@ -29,14 +30,14 @@ public class MaplePet extends Item {
     private Point pos;
     private int stance;
 
-    private MaplePet(int id, byte position, int uniqueid) {
+    private Pet(int id, byte position, int uniqueid) {
         super(id, position, (short) 1);
         this.uniqueid = uniqueid;
     }
 
-    public static MaplePet loadFromDb(int itemid, byte position, int petid) {
+    public static Pet loadFromDb(int itemid, byte position, int petid) {
         try {
-            MaplePet ret = new MaplePet(itemid, position, petid);
+            Pet ret = new Pet(itemid, position, petid);
             PreparedStatement ps = DatabaseConnection.getConnection().prepareStatement("SELECT * FROM pets WHERE petid = ?"); // Get pet details..
             ps.setInt(1, petid);
             ResultSet rs = ps.executeQuery();
@@ -49,23 +50,8 @@ public class MaplePet extends Item {
             ps.close();
             return ret;
         } catch (SQLException ex) {
-            Logger.getLogger(MaplePet.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(Pet.class.getName()).log(Level.SEVERE, null, ex);
             return null;
-        }
-    }
-
-    public void saveToDb() {
-        try {
-            PreparedStatement ps = DatabaseConnection.getConnection().prepareStatement("UPDATE pets SET name = ?, level = ?, closeness = ?, fullness = ? WHERE petid = ?");
-            ps.setString(1, getName()); // Set name
-            ps.setInt(2, getLevel()); // Set Level
-            ps.setInt(3, getCloseness()); // Set Closeness
-            ps.setInt(4, getFullness()); // Set Fullness
-            ps.setInt(5, getUniqueId()); // Set ID
-            ps.executeUpdate(); // Execute statement
-            ps.close();
-        } catch (SQLException ex) {
-            Logger.getLogger(MaplePet.class.getName()).log(Level.SEVERE, null, ex);
         }
     }
 
@@ -83,21 +69,42 @@ public class MaplePet extends Item {
             ps.close();
             return rs.getInt(1);
         } catch (SQLException ex) {
-            Logger.getLogger(MaplePet.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(Pet.class.getName()).log(Level.SEVERE, null, ex);
             return -1;
         }
     }
 
     public static void deletePet(int itemid, MapleClient c) {
+        Optional<IItem> petItem = c.getPlayer().getInventory(MapleInventoryType.CASH).findById(itemid);
+        if (petItem.isEmpty()) {
+            Logger.getLogger(Pet.class.getName()).log(Level.SEVERE, String.format("Cannot delete pet %d for character %d because the item cannot be found in their inventory.", itemid, c.getPlayer().getId()));
+            return;
+        }
+
         try {
             PreparedStatement ps = DatabaseConnection.getConnection().prepareStatement("DELETE FROM pets WHERE `petid` = ?");
-            ps.setInt(1, c.getPlayer().getInventory(MapleInventoryType.CASH).findById(itemid).getPetId());
+            ps.setInt(1, petItem.get().getPetId());
             ps.executeUpdate();
             ps.close();
         } catch (SQLException ex) {
-            Logger.getLogger(MaplePet.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(Pet.class.getName()).log(Level.SEVERE, null, ex);
         }
 
+    }
+
+    public void saveToDb() {
+        try {
+            PreparedStatement ps = DatabaseConnection.getConnection().prepareStatement("UPDATE pets SET name = ?, level = ?, closeness = ?, fullness = ? WHERE petid = ?");
+            ps.setString(1, getName()); // Set name
+            ps.setInt(2, getLevel()); // Set Level
+            ps.setInt(3, getCloseness()); // Set Closeness
+            ps.setInt(4, getFullness()); // Set Fullness
+            ps.setInt(5, getUniqueId()); // Set ID
+            ps.executeUpdate(); // Execute statement
+            ps.close();
+        } catch (SQLException ex) {
+            Logger.getLogger(Pet.class.getName()).log(Level.SEVERE, null, ex);
+        }
     }
 
     public String getName() {
@@ -164,20 +171,16 @@ public class MaplePet extends Item {
         this.stance = stance;
     }
 
-    public boolean canConsume(int itemId) {
-        for (int petId : MapleItemInformationProvider.getInstance().petsCanConsume(itemId)) {
-            if (petId == this.getItemId()) {
-                return true;
-            }
-        }
-        return false;
+    public boolean cannotConsume(int itemId) {
+        return MapleItemInformationProvider.getInstance().petsCanConsume(itemId).stream()
+                .noneMatch(id -> id == getItemId());
     }
 
     public void updatePosition(List<LifeMovementFragment> movement) {
         for (LifeMovementFragment move : movement) {
             if (move instanceof LifeMovement) {
                 if (move instanceof AbsoluteLifeMovement) {
-                    this.setPos(((LifeMovement) move).getPosition());
+                    this.setPos(move.getPosition());
                 }
                 this.setStance(((LifeMovement) move).getNewstate());
             }
